// src/utils/Enums/EHttpStatus.ts
export enum EHttpStatus {
	// Success
	OK = 200,
	CREATED = 201,
	ACCPETED = 202,
	NO_CONTENT = 204,
	RESET = 205,
	PARTIAL_CONTENT = 206,
	// Client Error
	BAD_REQUEST = 400,
	UNAUTHORIZED = 401,
	FORBIDDEN = 403,
	NOT_FOUND = 404,
	NOT_ALLOWED = 405,
	CONFLICT = 409,
	GONE = 410,
	LENGTH_REQUIRED = 411,
	PAYLOAD_TOO_LARGE = 413,
	UNSUPPORTED_MEDIA_TYPE = 415,
	TOO_MANY_REQUESTS = 429,
	// Server Error
	INTERNAL_SERVER_ERROR = 500,
	NOT_IMPLEMENTED = 501,
	SERVICE_UNAVAILABLE = 503
}